refactor: index -> Container 네이밍 통일
