+++ src/assets/common/svgs/index.ts
@@ -1,5 +1,5 @@
-export { default as File } from "./file.svg";
-export { default as Globe } from "./globe.svg";
-export { default as NextLogo } from "./next.svg";
-export { default as Vercel } from "./vercel.svg";
-export { default as Window } from "./window.svg";
+export { default as File } from './file.svg';
+export { default as Globe } from './globe.svg';
+export { default as NextLogo } from './next.svg';
+export { default as Vercel } from './vercel.svg';
+export { default as Window } from './window.svg';
