+++ src/components/pages/index.ts
@@ -1,2 +1,2 @@
-export { default as MainContainer } from './Main';
-export { default as RegisterContainer } from './Register';
+export { default as MainContainer } from './Main/Container';
+export { default as RegisterContainer } from './Register/Container';
